Adjust Directory
