+++ src/turbo/distance/neon/rotate/fht/fht.cc
@@ -20,7 +20,8 @@
 #include <cstdint>
 #include <cstring>
 #include <arm_neon.h>
-#include "scalar/fht/fht.h"
+#include "common/fht_common.h"
+#include "scalar/rotate/fht/fht.h"
 
 namespace zvec::turbo::neon {
 
@@ -117,101 +118,18 @@ void fht_vec_rescale_neon(float *data, size_t n, float factor) {
 
 void fht_rotate_neon(const float * /*in*/, float *out, size_t in_dim,
                      size_t /*out_dim*/, void *ctx) {
-  auto *base = reinterpret_cast<const uint8_t *>(ctx);
-  const size_t flip_offset = *reinterpret_cast<const size_t *>(base);
-  const size_t trunc_dim = *reinterpret_cast<const size_t *>(base + 8);
-  const float fac = *reinterpret_cast<const float *>(base + 16);
-  const uint8_t *flip = base + 24;
-  const size_t dim = in_dim;
-
-  if (trunc_dim == dim) {
-    fht_flip_sign_neon(flip, out, dim);
-    scalar::fht_inplace(out, trunc_dim);
-    fht_vec_rescale_neon(out, trunc_dim, fac);
-
-    fht_flip_sign_neon(flip + flip_offset, out, dim);
-    scalar::fht_inplace(out, trunc_dim);
-    fht_vec_rescale_neon(out, trunc_dim, fac);
-
-    fht_flip_sign_neon(flip + 2 * flip_offset, out, dim);
-    scalar::fht_inplace(out, trunc_dim);
-    fht_vec_rescale_neon(out, trunc_dim, fac);
-
-    fht_flip_sign_neon(flip + 3 * flip_offset, out, dim);
-    scalar::fht_inplace(out, trunc_dim);
-    fht_vec_rescale_neon(out, trunc_dim, fac);
-    return;
-  }
-
-  size_t start = dim - trunc_dim;
-  float *trunc_ptr = out + start;
-
-  fht_flip_sign_neon(flip, out, dim);
-  scalar::fht_inplace(out, trunc_dim);
-  fht_vec_rescale_neon(out, trunc_dim, fac);
-  fht_kacs_walk_neon(out, dim);
-
-  fht_flip_sign_neon(flip + flip_offset, out, dim);
-  scalar::fht_inplace(trunc_ptr, trunc_dim);
-  fht_vec_rescale_neon(trunc_ptr, trunc_dim, fac);
-  fht_kacs_walk_neon(out, dim);
-
-  fht_flip_sign_neon(flip + 2 * flip_offset, out, dim);
-  scalar::fht_inplace(out, trunc_dim);
-  fht_vec_rescale_neon(out, trunc_dim, fac);
-  fht_kacs_walk_neon(out, dim);
-
-  fht_flip_sign_neon(flip + 3 * flip_offset, out, dim);
-  scalar::fht_inplace(trunc_ptr, trunc_dim);
-  fht_vec_rescale_neon(trunc_ptr, trunc_dim, fac);
-  fht_kacs_walk_neon(out, dim);
-
-  fht_vec_rescale_neon(out, dim, 0.25f);
+  static constexpr FhtPrimitives kPrim = {
+      fht_flip_sign_neon, scalar::fht_inplace, fht_kacs_walk_neon,
+      fht_inv_kacs_walk_neon, fht_vec_rescale_neon};
+  fht_rotate_impl(out, in_dim, ctx, kPrim);
 }
 
 void fht_unrotate_neon(const float * /*in*/, float *out, size_t in_dim,
                        size_t /*out_dim*/, void *ctx) {
-  auto *base = reinterpret_cast<const uint8_t *>(ctx);
-  const size_t flip_offset = *reinterpret_cast<const size_t *>(base);
-  const size_t trunc_dim = *reinterpret_cast<const size_t *>(base + 8);
-  const float fac = *reinterpret_cast<const float *>(base + 16);
-  const uint8_t *flip = base + 24;
-  const size_t dim = in_dim;
-
-  if (trunc_dim == dim) {
-    for (int round = 3; round >= 0; --round) {
-      scalar::fht_inplace(out, trunc_dim);
-      fht_vec_rescale_neon(out, trunc_dim, fac);
-      fht_flip_sign_neon(flip + static_cast<size_t>(round) * flip_offset, out,
-                         dim);
-    }
-    return;
-  }
-
-  fht_vec_rescale_neon(out, dim, 4.0f);
-
-  size_t start = dim - trunc_dim;
-  float *trunc_ptr = out + start;
-
-  fht_inv_kacs_walk_neon(out, dim);
-  scalar::fht_inplace(trunc_ptr, trunc_dim);
-  fht_vec_rescale_neon(trunc_ptr, trunc_dim, fac);
-  fht_flip_sign_neon(flip + 3 * flip_offset, out, dim);
-
-  fht_inv_kacs_walk_neon(out, dim);
-  scalar::fht_inplace(out, trunc_dim);
-  fht_vec_rescale_neon(out, trunc_dim, fac);
-  fht_flip_sign_neon(flip + 2 * flip_offset, out, dim);
-
-  fht_inv_kacs_walk_neon(out, dim);
-  scalar::fht_inplace(trunc_ptr, trunc_dim);
-  fht_vec_rescale_neon(trunc_ptr, trunc_dim, fac);
-  fht_flip_sign_neon(flip + flip_offset, out, dim);
-
-  fht_inv_kacs_walk_neon(out, dim);
-  scalar::fht_inplace(out, trunc_dim);
-  fht_vec_rescale_neon(out, trunc_dim, fac);
-  fht_flip_sign_neon(flip, out, dim);
+  static constexpr FhtPrimitives kPrim = {
+      fht_flip_sign_neon, scalar::fht_inplace, fht_kacs_walk_neon,
+      fht_inv_kacs_walk_neon, fht_vec_rescale_neon};
+  fht_unrotate_impl(out, in_dim, ctx, kPrim);
 }
 
 }  // namespace zvec::turbo::neon
